Renames all directories and files to be snake case rather than no case and upper camel case for consistency
