+++ web/src/components/code_of_conduct/code_of_conduct.js
