+++ web/src/index.js
@@ -1,14 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { createHashRouter, RouterProvider } from "react-router-dom";
-import FeatureFlags from "./infra/FeatureFlags";
-import Contact from "./components/contact/Contact";
-import Schedule from "./components/schedule/Schedule";
-import About from "./components/about/About";
-import Health from "./components/health/Health";
-import CodeOfConduct from "./components/codeofconduct/CodeOfConduct";
-import Home from "./components/home/Home";
-import NavBar from "./components/navbar/NavBar";
+import FeatureFlags from "./infra/feature_flags";
+import Contact from "./components/contact/contact";
+import Schedule from "./components/schedule/schedule";
+import About from "./components/about/about";
+import Health from "./components/health/health";
+import CodeOfConduct from "./components/code_of_conduct/code_of_conduct";
+import Home from "./components/home/home";
+import NavBar from "./components/nav_bar/nav_bar";
 import ThemeProvider from "@mui/material/styles/ThemeProvider";
 import theme from "./common/theme";
 import "./index.css";
